Statring preventing seleting item to rebuild the whole page
