+++ lib/loan/ui/pages/loan_group_page/add_edit_loan_page.dart
@@ -2,23 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:myecl/loan/class/loan.dart';
-import 'package:myecl/loan/providers/admin_loan_list_provider.dart';
-import 'package:myecl/loan/providers/loaner_loan_list_provider.dart';
-import 'package:myecl/loan/providers/item_list_provider.dart';
 import 'package:myecl/loan/providers/loan_provider.dart';
-import 'package:myecl/loan/providers/loaner_provider.dart';
-import 'package:myecl/loan/providers/selected_items_provider.dart';
 import 'package:myecl/loan/class/item.dart';
-import 'package:myecl/loan/providers/loan_page_provider.dart';
 import 'package:myecl/loan/tools/constants.dart';
 import 'package:myecl/loan/tools/functions.dart';
+import 'package:myecl/loan/ui/pages/loan_group_page/add_edit_button.dart';
 import 'package:myecl/loan/ui/pages/loan_group_page/date_entry.dart';
 import 'package:myecl/loan/ui/pages/loan_group_page/item_bar.dart';
 import 'package:myecl/loan/ui/pages/loan_group_page/search_result.dart';
+import 'package:myecl/loan/ui/pages/loan_group_page/start_date_entry.dart';
 import 'package:myecl/loan/ui/text_entry.dart';
 import 'package:myecl/tools/functions.dart';
 import 'package:myecl/tools/token_expire_wrapper.dart';
-import 'package:myecl/tools/ui/shrink_button.dart';
 import 'package:myecl/user/providers/user_list_provider.dart';
 
 class AddEditLoanPage extends HookConsumerWidget {
@@ -26,14 +21,8 @@ class AddEditLoanPage extends HookConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final pageNotifier = ref.watch(loanPageProvider.notifier);
-    final adminLoanListNotifier = ref.watch(adminLoanListProvider.notifier);
     final key = GlobalKey<FormState>();
-    final items = ref.watch(itemListProvider);
-    final selectedItems = ref.watch(editSelectedListProvider);
-    final loanListNotifier = ref.watch(loanerLoanListProvider.notifier);
     final loan = ref.watch(loanProvider);
-    final loaner = ref.watch(loanerProvider);
     final isEdit = loan.id != Loan.empty().id;
     final borrower = useState(loan.borrower);
     final note = useTextEditingController(text: loan.notes);
@@ -49,10 +38,6 @@ class AddEditLoanPage extends HookConsumerWidget {
     final numberSelected = useState(loan.items.length);
     final initialDate = useState(isEdit ? loan.start : DateTime.now());
 
-    void displayToastWithContext(TypeMsg type, String msg) {
-      displayToast(context, type, msg);
-    }
-
     void evaluateEnd(List<Item> selected) {
       end.text = processDate(DateTime.parse(processDateBack(start.text)).add(
           Duration(
@@ -64,6 +49,7 @@ class AddEditLoanPage extends HookConsumerWidget {
                           : previousValue)).toInt())));
     }
 
+    print("AddEditLoanPage.build");
     return SingleChildScrollView(
         physics: const BouncingScrollPhysics(),
         child: Form(
@@ -187,26 +173,11 @@ class AddEditLoanPage extends HookConsumerWidget {
                 SearchResult(
                     borrower: borrower, queryController: queryController),
                 const SizedBox(height: 30),
-                DateEntry(
-                  title: LoanTextConstants.beginDate,
-                  controller: start,
-                  dateBefore: DateTime.now(),
-                  onSelect: () {
-                    items.whenData((itemList) {
-                      List<Item> selected = itemList
-                          .where((element) =>
-                              selectedItems[itemList.indexOf(element)])
-                          .toList();
-                      if (selected.isNotEmpty) {
-                        evaluateEnd(selected);
-                      } else {
-                        end.text = "";
-                      }
-                      initialDate.value =
-                          DateTime.parse(processDateBack(start.text));
-                    });
-                  },
-                ),
+                StartDateEntry(
+                    end: end,
+                    start: start,
+                    initialDate: initialDate,
+                    evaluateEnd: evaluateEnd),
                 const SizedBox(height: 30),
                 DateEntry(
                   title: LoanTextConstants.endDate,
@@ -231,137 +202,25 @@ class AddEditLoanPage extends HookConsumerWidget {
                   suffix: '',
                 ),
                 const SizedBox(height: 50),
-                ShrinkButton(
-                  waitChild: Container(
-                      width: double.infinity,
-                      padding: const EdgeInsets.only(top: 8, bottom: 12),
-                      alignment: Alignment.center,
-                      decoration: BoxDecoration(
-                        color: Colors.black,
-                        borderRadius: BorderRadius.circular(10),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.grey.withOpacity(0.5),
-                            spreadRadius: 5,
-                            blurRadius: 10,
-                            offset: const Offset(3, 3),
-                          ),
-                        ],
-                      ),
-                      child: const Center(
-                        child: SizedBox(
-                          height: 25,
-                          width: 25,
-                          child: CircularProgressIndicator(
-                              valueColor:
-                                  AlwaysStoppedAnimation<Color>(Colors.white)),
-                        ),
-                      )),
-                  onTap: () async {
+                AddEditButton(
+                  isEdit: isEdit,
+                  borrower: borrower,
+                  note: note,
+                  start: start,
+                  end: end,
+                  caution: caution,
+                  onAddEdit: (p0) async {
                     if (key.currentState == null) {
                       return;
                     }
                     if (key.currentState!.validate() &&
                         borrower.value.id.isNotEmpty) {
-                      if (processDateBack(start.text)
-                              .compareTo(processDateBack(end.text)) >=
-                          0) {
-                        displayToast(context, TypeMsg.error,
-                            LoanTextConstants.invalidDates);
-                      } else {
-                        await items.when(
-                          data: (itemList) async {
-                            await tokenExpireWrapper(ref, () async {
-                              List<Item> selected = itemList
-                                  .where((element) =>
-                                      selectedItems[itemList.indexOf(element)])
-                                  .toList();
-                              if (selected.isNotEmpty) {
-                                Loan newLoan = Loan(
-                                  loaner: isEdit ? loan.loaner : loaner,
-                                  items: selected,
-                                  borrower: borrower.value,
-                                  caution: caution.text.isNotEmpty
-                                      ? caution.text
-                                      : "${selected.fold<double>(0, (previousValue, element) => previousValue + element.caution)}€",
-                                  end:
-                                      DateTime.parse(processDateBack(end.text)),
-                                  id: isEdit ? loan.id : "",
-                                  notes: note.text,
-                                  start: DateTime.parse(
-                                      processDateBack(start.text)),
-                                  returned: false,
-                                );
-                                final value = isEdit
-                                    ? await loanListNotifier.updateLoan(newLoan)
-                                    : await loanListNotifier.addLoan(newLoan);
-                                if (value) {
-                                  await adminLoanListNotifier.setTData(
-                                      isEdit ? loan.loaner : loaner,
-                                      await loanListNotifier.copy());
-                                  pageNotifier.setLoanPage(LoanPage.admin);
-                                  if (isEdit) {
-                                    displayToastWithContext(TypeMsg.msg,
-                                        LoanTextConstants.updatedLoan);
-                                  } else {
-                                    displayToastWithContext(TypeMsg.msg,
-                                        LoanTextConstants.addedLoan);
-                                  }
-                                } else {
-                                  if (isEdit) {
-                                    displayToastWithContext(TypeMsg.error,
-                                        LoanTextConstants.updatingError);
-                                  } else {
-                                    displayToastWithContext(TypeMsg.error,
-                                        LoanTextConstants.addingError);
-                                  }
-                                }
-                              } else {
-                                displayToastWithContext(TypeMsg.error,
-                                    LoanTextConstants.noItemSelected);
-                              }
-                            });
-                          },
-                          error: (error, s) {
-                            displayToast(
-                                context, TypeMsg.error, error.toString());
-                          },
-                          loading: () {
-                            displayToast(context, TypeMsg.error,
-                                LoanTextConstants.addingError);
-                          },
-                        );
-                      }
+                      p0();
                     } else {
                       displayToast(context, TypeMsg.error,
                           LoanTextConstants.incorrectOrMissingFields);
                     }
                   },
-                  child: Container(
-                      width: double.infinity,
-                      padding: const EdgeInsets.only(top: 8, bottom: 12),
-                      alignment: Alignment.center,
-                      decoration: BoxDecoration(
-                        color: Colors.black,
-                        borderRadius: BorderRadius.circular(10),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.grey.withOpacity(0.5),
-                            spreadRadius: 5,
-                            blurRadius: 10,
-                            offset: const Offset(
-                                3, 3), // changes position of shadow
-                          ),
-                        ],
-                      ),
-                      child: Text(
-                          isEdit
-                              ? LoanTextConstants.edit
-                              : LoanTextConstants.add,
-                          style: const TextStyle(
-                              color: Colors.white,
-                              fontSize: 25,
-                              fontWeight: FontWeight.bold))),
                 ),
                 const SizedBox(height: 30),
               ]),
